+++ EatPic-iOS/Sources/Models/Record/HashtagRecordModel.swift
@@ -153,411 +153,3 @@ extension HashtagRecordModel {
         )
     }
 }
-
-protocol HashtagTitlePolicy {
-    // 타이틀이 허용되는지 검증( 성공: ok, 실패: 에러 반환)
-    func validate(_ title: String) -> Result<Void, HashtagPolicyError>
-    
-    var maxHangul: Int { get }
-}
-
-enum HashtagPolicyError: LocalizedError {
-    case tooLongHangul(limit: Int)
-    case empty
-    case duplicateTitle
-    // 추후 확장
-    
-    var errorDescription: String? {
-        switch self {
-        case .empty: return "내용을 입력해 주세요."
-        case .tooLongHangul(let limit):
-            return "최대 \(limit)글자까지 입력 가능합니다."
-        case .duplicateTitle:
-            return "이미 존재하는 해시태그입니다."
-        }
-    }
-}
-
-struct DefaultHashtagTitlePolicy: HashtagTitlePolicy {
-    let maxHangul: Int = 5
-    
-    func validate(_ title: String) -> Result<Void, HashtagPolicyError> {
-        let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard trimmed.isEmpty == false else { return .failure(.empty) }
-        
-        let length = trimmed.count
-        
-        return length <= maxHangul ? .success(()) : .failure(.tooLongHangul(limit: maxHangul))
-    }
-}
-
-@Observable
-@MainActor
-final class HashtagRecordViewModel {
-    private(set) var model: HashtagRecordModel
-    private(set) var catalog: HashtagCatalogProviding
-    private let hashtagTitlePolicy: HashtagTitlePolicy
-    
-    private(set) var customAddedSinceReset: [HashtagCategory] = []
-    
-    init(
-        date: Date = .now,
-        catalog: HashtagCatalogProviding = DefaultHashtagCatalog(),
-        hashtagTitlePolicy: HashtagTitlePolicy = DefaultHashtagTitlePolicy()
-    ) {
-        self.catalog = catalog
-        self.model = .initial(for: date, catalog: catalog)
-        self.hashtagTitlePolicy = hashtagTitlePolicy
-    }
-    
-    var date: Date { model.date }
-    var selectedHashtags: [HashtagCategory]? { model.selectedHashtags }
-    var selectedCount: Int { model.selectedCount }
-    var canSelectMore: Bool { model.canSelectMore }
-    var isFull: Bool { model.isFull }
-    var checks: [HashtagCheck] { model.checks }
-    
-    // MARK: - 조회
-    
-    func isSelected(_ id: HashtagCategory.ID) -> Bool {
-        model.check(of: id)?.selected == true
-    }
-    
-    func selectedAt(_ id: HashtagCategory.ID) -> Date? {
-        model.check(of: id)?.selectedAt
-    }
-    
-    // MARK: - Domain Mutation (사이드 이팩트 free)
-    
-    func hashtagToggle(_ id: HashtagCategory.ID, at time: Date = .now) {
-        model.toggle(id, at: time)
-    }
-    
-    func markUploaded(_ id: HashtagCategory.ID, at time: Date = .now) {
-        model.markUploaded(id, at: time)
-    }
-    
-    func reset(for date: Date = .now) {
-        model = .initial(for: date, catalog: catalog)
-        customAddedSinceReset.removeAll()
-    }
-    
-    // MARK: Runtime extension (사용자에 의한 태그 추가)
-    
-    /**
-     사용자 정의 해시태그를 모델에 추가합니다.
-     
-     - Parameters:
-     - id: 해시태그 고유 식별자(서버/영속 키). `nil`인 경우 `title`을 가공하여 생성합니다.
-     - title: 사용자에게 표시되는 해시태그 제목.
-     - selectImmediately: `true`면 추가 직후 해당 해시태그를 선택 상태로 전환합니다. 기본값은 `false`입니다.
-     
-     - Returns:
-     - `.success(HashtagCategory)`: 정상적으로 추가된 해시태그 카테고리.
-     - `.failure(HashtagPolicyError)`: 제목이 비어 있거나, 길이 제한 등 정책을 위반한 경우.
-     
-     - Note:
-     - 이 메서드는 해시태그 목록(`checks`)에 동일한 `id`가 이미 존재하면 아무 동작도 하지 않습니다.
-     - 최대 선택 개수 제한(예: 3개)은 `selectImmediately` 시점의 토글 동작에 의해 적용됩니다.
-     */
-    @discardableResult
-    func addCustomCategory(
-        id: String? = nil,
-        title: String,
-        selectImmediately: Bool = false
-    ) -> Result<HashtagCategory, HashtagPolicyError> {
-        switch hashtagTitlePolicy.validate(title) {
-        case .success:
-            guard isDuplicateTitle(title) == false else {
-                return .failure(.duplicateTitle)
-            }
-            
-            let key = id ?? title.trimmingCharacters(in: .whitespacesAndNewlines)
-                .replacingOccurrences(of: " ", with: "_")
-            
-            let newHashtag = HashtagCategory(id: key, title: title)
-            model.appendHashtagIfNeeded(newHashtag)
-            customAddedSinceReset.append(newHashtag)
-            
-            if selectImmediately { model.toggle(key) }
-            return .success(newHashtag)
-        case .failure(let err):
-            return .failure(err)
-        }
-    }
-    
-    /// 입력된 title로 사용자 태그를 추가할 수 있는지 여부를 검증합니다.
-    /// - Returns: 가능하면 `true`, 불가하면 `false`
-    func canAdd(_ title: String) -> Bool {
-        switch hashtagTitlePolicy.validate(title) {
-        case .success: return !isDuplicateTitle(title)
-        case .failure: return false
-        }
-    }
-    
-    private func normalized(_ title: String) -> String {
-        title.trimmingCharacters(in: .whitespacesAndNewlines)
-            .folding(options: [.diacriticInsensitive, .caseInsensitive], locale: .current)
-             .lowercased()
-    }
-
-    private func isDuplicateTitle(_ title: String) -> Bool {
-        let key = normalized(title)
-        let existing = model.checks.map { normalized($0.hashtag.title) }
-        return existing.contains(key)
-    }
-}
-
-import SwiftUI
-
-typealias HashtagRecordVMFactory = @MainActor (_ date: Date) -> HashtagRecordViewModel
-
-private struct AlertPayload: Identifiable {
-    let id = UUID()
-    let title: String
-    let message: String
-}
-
-struct HashtagSelectingView: View {
-    @EnvironmentObject private var container: DIContainer
-    @EnvironmentObject private var recordFlowVM: RecordFlowViewModel
-    
-    @State private var hashtagRecordVM: HashtagRecordViewModel
-    
-    @State private var alert: AlertPayload?
-    @State private var addHashtagSheet: Bool = false
-    @State private var addHashtagText: String = ""
-    
-    private let columns = [
-        GridItem(.adaptive(minimum: 50)),
-        GridItem(.adaptive(minimum: 50)),
-        GridItem(.adaptive(minimum: 50)),
-        GridItem(.adaptive(minimum: 50))
-    ]
-    
-    init(
-        date: Date = .now,
-        factory: @escaping HashtagRecordVMFactory = { date in
-            HashtagRecordViewModel(date: date)
-        }
-    ) {
-        self._hashtagRecordVM = State(wrappedValue: factory(date))
-    }
-
-    var body: some View {
-        ZStack {
-            HastagSelectingContentView(
-                hashtagRecordVM: hashtagRecordVM,
-                alert: $alert,
-                addHashtagSheet: $addHashtagSheet,
-                addHashtagText: $addHashtagText
-            )
-            
-            if let alert {
-                AlertModalView(
-                    messageTitle: alert.title,
-                    messageDescription: alert.message,
-                    messageColor: .black,
-                    btnText: "확인",
-                    btnAction: {
-                        self.alert = nil
-                    }
-                )
-            }
-        }
-    }
-}
-
-private struct HastagSelectingContentView: View {
-    @EnvironmentObject private var container: DIContainer
-    @EnvironmentObject private var recordFlowVM: RecordFlowViewModel
-    @Bindable private var hashtagRecordVM: HashtagRecordViewModel
-    
-    @Binding private var alert: AlertPayload?
-    @Binding private var addHashtagSheet: Bool
-    @Binding private var addHashtagText: String
-    
-    init(
-        hashtagRecordVM: HashtagRecordViewModel,
-        alert: Binding<AlertPayload?>,
-        addHashtagSheet: Binding<Bool>,
-        addHashtagText: Binding<String>
-    ) {
-        self.hashtagRecordVM = hashtagRecordVM
-        self._alert = alert
-        self._addHashtagSheet = addHashtagSheet
-        self._addHashtagText = addHashtagText
-    }
-    
-    var body: some View {
-        VStack {
-            // 상단 이미지
-            Image("Record/img_record_itcong")
-                .resizable()
-                .frame(width: 180, height: 180)
-
-            Spacer().frame(height: 36)
-
-            // 안내 텍스트
-            Text("식사와 연관된\n해시태그를 선택해주세요")
-                .font(.dsTitle2)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            Spacer().frame(height: 8)
-
-            Text("최대 3개를 선택할 수 있어요")
-                .font(.dsFootnote)
-                .foregroundStyle(Color.gray060)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            Spacer().frame(height: 32)
-
-            // 해시태그 버튼 영역
-            hastagBtnList()
-            
-            Spacer().frame(height: 80)
-
-            addHashTagBtn
-
-            Spacer().frame(height: 51)
-
-            // 하단 확인 버튼
-            PrimaryButton(
-                color: .green060,
-                text: "확인",
-                font: .dsTitle3,
-                textColor: .white,
-                width: 361,
-                height: 48,
-                cornerRadius: 10
-            ) {
-                guard let selectedHashtags = hashtagRecordVM.selectedHashtags else {
-                    withAnimation {
-                        alert = .init(title: "안내", message: "하나 이상의 해시태그를 선택해 주세요.")
-                    }
-                    return
-                }
-                recordFlowVM.setTags(selectedHashtags)
-                if recordFlowVM.canProceedToRecord {
-                    container.router.push(.picCardRecord)
-                }
-            }
-        }
-        .padding(.horizontal, 16)
-        .sheet(isPresented: $addHashtagSheet) {
-            BottomSheetView(title: "해시태그 추가") {
-                HashtagAddView(
-                    hashtagInput: $addHashtagText,
-                    isEnabled: .constant(hashtagRecordVM.canAdd(addHashtagText))
-                ) {
-                    let result = hashtagRecordVM.addCustomCategory(title: addHashtagText)
-                    switch result {
-                    case .success:
-                        break
-                    case .failure(let err):
-                        withAnimation {
-                            alert = .init(title: "안내", message: err.localizedDescription)
-                        }
-                    }
-                    addHashtagSheet = false
-                    addHashtagText = ""
-                }
-            }
-            .presentationDetents([.height(200)])
-        }
-        .customNavigationBar {
-            Text("Pic 카드 기록")
-                .font(.dsTitle2)
-                .foregroundStyle(Color.gray080)
-        } right: {
-            Button(action: {
-                container.router.popToRoot()
-            }, label: {
-                Image("Record/btn_home_close")
-            })
-        }
-    }
-    
-    private func hastagBtnList(chunkSize: Int = 4) -> some View {
-        let rows = hashtagRecordVM.checks.map(\.hashtag).chunked(into: chunkSize)
-        
-        return LazyVStack(alignment: .leading, spacing: 8) {
-            ForEach(Array(rows.enumerated()), id: \.offset) { _, row in
-                HStack(alignment: .top, spacing: 8) {
-                    ForEach(row, id: \.self) { hashtag in
-                        hashtagBtn(hashtag: hashtag)
-                    }
-                    
-                    Spacer() // 남는 공간은 우측으로 밀어 좌측정렬 유지
-                }
-            }
-        }
-    }
-    
-    private func hashtagBtn(hashtag: HashtagCategory) -> some View {
-        let selected = hashtagRecordVM.isSelected(hashtag.id)
-        
-        return Button {
-            hashtagRecordVM.hashtagToggle(hashtag.id)
-        } label: {
-            Text("#\(hashtag.title)")
-                .font(.dsCallout)
-                .foregroundStyle(
-                    selected ? Color.green060: .gray050)
-                .padding(.horizontal, 14)
-                .padding(.vertical, 10)
-                .background(
-                    selected ? Color.green010: Color.white)
-                .clipShape(RoundedRectangle(cornerRadius: 50))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 50)
-                        .stroke(
-                            selected ? Color.green060: Color.gray050,
-                            lineWidth: 1)
-                )
-        }
-        .buttonStyle(.plain)
-        .disabled(!selected && !hashtagRecordVM.canSelectMore) // 3개 이상 선택시 나머지 비활성화
-    }
-    
-    private var addHashTagBtn: some View {
-        HStack {
-            Spacer()
-            
-            Button {
-                addHashtagSheet.toggle()
-            } label: {
-                HStack {
-                    Text("직접 추가하기")
-                        .font(.dsSubhead)
-                        .foregroundStyle(Color.gray060)
-                    Spacer()
-                    Image("Record/btn_record_add")
-                }
-                .padding(.horizontal, 10)
-                .padding(.vertical, 8)
-                .frame(width: 135, height: 38)
-                .background(Color.gray020)
-                .clipShape(RoundedRectangle(cornerRadius: 10))
-            }
-            .buttonStyle(.plain)
-
-            Spacer()
-        }
-    }
-}
-
-private extension Array {
-    func chunked(into size: Int) -> [[Element]] {
-        guard size > 0 else { return [self] }
-        return stride(from: 0, to: count, by: size).map {
-            Array(self[$0 ..< Swift.min($0 + size, count)])
-        }
-    }
-}
-
-#Preview {
-    HashtagSelectingView()
-        .environmentObject(RecordFlowViewModel())
-        .environmentObject(DIContainer())
-}
